requirements.txt
fastapi 
uvicorn 
SQLAlchemy
python-decouple
asyncpg
psycopg2-binary
aiogram
python-dotenv
APScheduler
pydantic
openpyxl
